requirements.txt
fastapi
asgiref
werkzeug
openai
python-dotenv
